US-BE-09 Admin Functionality
Admins are added to the app! The creator of the group is an admin by default. When the last admin of the group leaves it, no admins are left in the group (unlike Whatsapp since a remaining user is given admin rights). An admin can edit the group information, change the permission of other participants and expel them from the group.